Move `HaghsDataUpdateCoordinator` from `__init__.py` to a new file `coordinator.py`.
